feat(window-controls): refactor window control functions for clarity and consistency
